requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
arxiv==2.1.3
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backports.tarfile==1.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
command-not-found==0.3
cryptography==43.0.0
dataclasses-json==0.6.7
dbus-python==1.2.18
distro==1.9.0
distro-info==1.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
feedparser==6.0.11
Flask==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.3.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.29
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langgraph==0.2.3
langgraph-checkpoint==1.0.2
langsmith==0.1.98
launchpadlib==2.0.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.3
more-itertools==10.4.0
multidict==6.0.5
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==2.0.1
oauthlib==3.2.2
openai==1.40.1
orjson==3.10.6
packaging==24.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.21.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.9.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-apt==2.4.0+ubuntu1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
SecretStorage==3.3.3
sgmllib3k==1.0.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
systemd-python==234
tenacity==9.0.0
tiktoken==0.7.0
tqdm==4.66.5
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
wadllib==1.3.6
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
